requirements.txt
torch>=2.0.0
transformers>=4.35.0
datasets>=2.14.0
accelerate>=0.24.0
tensorboard>=2.14.0
wandb>=0.15.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
tqdm>=4.66.0
matplotlib>=3.7.0
seaborn>=0.12.0
pyyaml>=6.0
pytest>=7.4.0
jupyter>=1.0.0
ipykernel>=6.25.0
